查勘提供关联多权证

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/BasicApplyMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BasicApplyMapper {
-    int countByExample(BasicApplyExample example);
+    long countByExample(BasicApplyExample example);
 
     int deleteByExample(BasicApplyExample example);
